devtools/client/performance-new/test/chrome/test_perf-state-01.html

<!DOCTYPE HTML>
<html>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<head>
  <meta charset="utf-8">
  <title>Perf component test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
  <div id="container"></div>

  <pre id="test">
    <script src="head.js" type="application/javascript"></script>
    <script type="application/javascript">
      "use strict";

      /**
       * Test the normal workflow of starting and stopping the profiler through the
       * Perf component.
       */
      addPerfTest(async () => {
        const {
          perfFrontMock,
          getRecordingState,
          receiveProfileCalls,
          mountComponent,
          container,
        } = createPerfComponent();

        mountComponent();

        is(getRecordingState(), "not-yet-known",
          "The component at first is in an unknown state.");

        await perfFrontMock._flushAsyncQueue();
        is(getRecordingState(), "available-to-record",
          "After talking to the actor, we're ready to record.");

        const button = container.querySelector("button");
        ok(button, "Selected the button to click.");
        button.click();
        is(getRecordingState(), "request-to-start-recording",
          "Sent in a request to start recording.");

        await perfFrontMock._flushAsyncQueue();
        is(getRecordingState(), "recording",
          "The actor has started its recording");

        button.click();
        is(getRecordingState(),
          "request-to-get-profile-and-stop-profiler",
          "We have requested to stop the profiler.");

        await perfFrontMock._flushAsyncQueue();
        is(getRecordingState(), "available-to-record",
          "The profiler is available to record again.");
        await perfFrontMock._flushAsyncQueue();
        is(receiveProfileCalls.length, 1,
           "The receiveProfile function was called once");
        is(typeof receiveProfileCalls[0], "object", "Got a profile");
      });
    </script>
  </pre>
</body>
</html>
